Create and sort each class into its proper package
Previously, all classes were in the same file and thus in the same package, leading to messier coding.

Created new folders and packages and sorted each class into its appropriate package.
