+++ src/test/java/puke/PukeTest.java
@@ -1,6 +1,7 @@
 package puke;
 
 import org.junit.jupiter.api.Test;
+import puke.command.*;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
